+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `highlight-own-exchange-orders`: Make own order rows bold
+- `item-icons`: Add a detail to the INS icon
 - Swap the order of SF and FF in category sorting
 
 ## 25.4.27