+++ CHANGELOG.md
@@ -14,7 +14,9 @@
 ### Fixed
 
 - `XIT ACT`: Fix CX Buy action using stale order book data
+- `cxob-depth-bars`: Fix feature not working for newly placed orders
 - `cxpo-order-book`: Fix price/quantity autofill number formatting
+- `highlight-own-exchange-orders`: Fix feature not working for newly placed orders
 
 ## 25.4.27