+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - `other-context-notification-count`: (new) Displays the number of notifications from other contexts in the NOTS header label
 - Add user data backups (up to 5, every 24 hours)
-- Add user data restore from backup after extension reinstall
+- Add user data restore from backup after the extension reinstall
 
 ### Changed